processes/requirements.txt
kafka-python==1.3.3
psycopg2
google-api-python-client